+++ qa-admin/src/main/client/src/App.tsx
@@ -1,5 +1,6 @@
 import {BrowserRouter as Router} from "react-router-dom"
 import {AppRouter} from "./components/AppRouter";
+import "./http/apiResInterceptors"
 
 function App() {
     return (
